--- tests/Foundation/Fixtures/EventDiscovery/Events/EventTwo.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Illuminate\Tests\Foundation\Fixtures\EventDiscovery\Events;
-
-class EventTwo
-{
-}
