--- tests/Foundation/Fixtures/EventDiscovery/Listeners/Listener.php
@@ -1,24 +1,0 @@
-<?php
-
-namespace Illuminate\Tests\Foundation\Fixtures\EventDiscovery\Listeners;
-
-use Illuminate\Tests\Foundation\Fixtures\EventDiscovery\Events\EventOne;
-use Illuminate\Tests\Foundation\Fixtures\EventDiscovery\Events\EventTwo;
-
-class Listener
-{
-    public function handle(EventOne $event)
-    {
-        //
-    }
-
-    public function handleEventOne(EventOne $event)
-    {
-        //
-    }
-
-    public function handleEventTwo(EventTwo $event)
-    {
-        //
-    }
-}
