+++ tests/Integration/Foundation/Fixtures/EventDiscovery/Events/EventOne.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Illuminate\Tests\Integration\Foundation\Fixtures\EventDiscovery\Events;
+
+class EventOne
+{
+}
