+++ tests/Integration/Foundation/Fixtures/EventDiscovery/Events/EventTwo.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Illuminate\Tests\Integration\Foundation\Fixtures\EventDiscovery\Events;
+
+class EventTwo
+{
+}
