+++ tests/Integration/Foundation/Fixtures/EventDiscovery/Listeners/Listener.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace Illuminate\Tests\Integration\Foundation\Fixtures\EventDiscovery\Listeners;
+
+use Illuminate\Tests\Integration\Foundation\Fixtures\EventDiscovery\Events\EventOne;
+use Illuminate\Tests\Integration\Foundation\Fixtures\EventDiscovery\Events\EventTwo;
+
+class Listener
+{
+    public function handle(EventOne $event)
+    {
+        //
+    }
+
+    public function handleEventOne(EventOne $event)
+    {
+        //
+    }
+
+    public function handleEventTwo(EventTwo $event)
+    {
+        //
+    }
+}
